fix conversation
