--- src/main/java/com/bathanh/exercise/basic/array/Ex304.java
@@ -1,15 +1,0 @@
-package com.bathanh.exercise.basic.array;
-
-public class Ex304 {
-
-    public boolean checkRepeatString(final String s) {
-        for (int i = 0; i < s.length() - 1; i++) {
-            for (int j = i + 1; j < s.length(); j++) {
-                if (s.charAt(i) == s.charAt(j)) {
-                    return true;
-                }
-            }
-        }
-        return false;
-    }
-}
